@@ -16,31 +16,31 @@
     "@types/react": "*"
   },
   "dependencies": {
-    "core-native": "file:../../core-native-project/build/dist",
-    "native-base": "2.12.1",
-    "react": "16.8.6",
-    "react-native": "0.60.3",
-    "react-native-device-info": "2.2.2",
-    "react-native-gesture-handler": "1.3.0",
-    "react-navigation": "3.11.1"
+    "core-native": "1.6.2",
+    "native-base": "2.13.8",
+    "react": "16.9.0",
+    "react-native": "0.61.5",
+    "react-native-device-info": "5.4.0",
+    "react-native-gesture-handler": "1.5.2",
+    "react-navigation": "4.0.10"
   },
   "devDependencies": {
-    "@babel/plugin-proposal-decorators": "7.4.4",
-    "@babel/plugin-transform-flow-strip-types": "7.4.4",
-    "@babel/runtime": "7.5.4",
-    "@types/jest": "24.0.15",
+    "@babel/plugin-proposal-decorators": "7.7.4",
+    "@babel/plugin-transform-flow-strip-types": "7.7.4",
+    "@babel/runtime": "7.7.7",
+    "@types/jest": "24.0.25",
     "axios": "0.19.0",
-    "babel-plugin-module-resolver": "3.2.0",
+    "babel-plugin-module-resolver": "4.0.0",
     "babel-plugin-transform-remove-console": "6.9.4",
-    "chalk": "2.4.2",
+    "chalk": "3.0.0",
     "fs-extra": "8.1.0",
-    "jest": "24.8.0",
-    "metro-react-native-babel-preset": "0.55.0",
-    "prettier": "1.18.2",
-    "ts-jest": "24.0.2",
-    "tslint": "5.18.0",
-    "tslint-react": "4.0.0",
-    "typescript": "3.5.3",
-    "yargs": "13.2.4"
+    "jest": "24.9.0",
+    "metro-react-native-babel-preset": "0.57.0",
+    "prettier": "1.19.1",
+    "ts-jest": "24.2.0",
+    "tslint": "5.20.1",
+    "tslint-react": "4.1.0",
+    "typescript": "3.7.4",
+    "yargs": "15.0.2"
   }
 }
